Revert "COMAAS-459 Fix unread messages counter in Cassandra - Removing CounterIncrementController, legacy functionality regarding only ebayk - Cassandra counter is not re-written by 0 or 1 if the conversation is updated but should update real number of unread mes"
This reverts commit ff5649f27f952cf57cc16d54b3547525d5c572aa.

Reviewed-on: https://gerrit.ecg.so/59335
Tested-by: Gerrit-At-Gitmaster001

+++ messagecenters/core-messagecenter/src/test/java/com/ecg/messagecenter/persistence/simple/PostBoxTest.java
@@ -56,7 +56,7 @@ public class PostBoxTest {
         ), 180);
 
         postBox.markConversationUnread("convid2", null);
-        assertEquals(1, postBox.getNewRepliesCounter().getValue());
+        assertEquals(1l, postBox.getNewRepliesCounter().getValue().longValue());
         assertTrue(postBox.getUnreadConversations().containsKey("convid2"));
     }
